fix-Export-linked-file-using-api
