+++ Source/IFCExporterUIOverride/IFCCommandOverrideApplication.cs
@@ -289,56 +289,11 @@ namespace BIM.IFC.Export.UI
                   selectedConfig.ActiveViewId = selectedConfig.UseActiveViewGeometry ? activeViewId : ElementId.InvalidElementId;
                   selectedConfig.UpdateOptions(exportOptions, activeViewId);
 
-                  IDictionary<ElementId, string> linkGUIDsCache =
-                     new Dictionary<ElementId, string>();
-
-                  IDictionary<RevitLinkInstance, Transform> linkInstanceTranforms = null;
-                  string errorMessage = null;
-                  int numBadInstances = 0;
-
+                  IFCLinkedDocumentExporter linkExporter = null;
                   if (exportLinks)
                   {
-                     // Cache for links guids
-                     FilteredElementCollector collector = new FilteredElementCollector(document);
-                     ElementFilter elementFilter = new ElementClassFilter(typeof(RevitLinkInstance));
-                     List<RevitLinkInstance> rvtLinkInstances =
-                        collector.WherePasses(elementFilter).Cast<RevitLinkInstance>().ToList();
-
-                     string linkGuidOptionString = string.Empty;
-
-                     // Get the transforms of the links we can export, and error message information,
-                     // including number, for the ones we can't.
-                     (linkInstanceTranforms, errorMessage, numBadInstances) =
-                        GetLinkedInstanceInfo(rvtLinkInstances);
-
-                     ISet<string> existingGUIDs = new HashSet<string>();
-
-                     foreach (RevitLinkInstance linkInstance in linkInstanceTranforms.Keys)
-                     {
-                        Parameter parameter = linkInstance.get_Parameter(BuiltInParameter.IFC_GUID);
-
-                        string sGUID = GUIDUtil.GetSimpleElementIFCGUID(linkInstance);
-
-                        string sGUIDlower = sGUID.ToLower();
-
-                        while (existingGUIDs.Contains(sGUIDlower))
-                        {
-                           sGUID += "-";
-                           sGUIDlower += "-";
-                        }
-                        existingGUIDs.Add(sGUIDlower);
-
-                        if (doFederatedExport)
-                           linkGuidOptionString += linkInstance.Id.ToString() + "," + sGUID + ";";
-                        else
-                           linkGUIDsCache.Add(linkInstance.Id, sGUID);
-                     }
-
-                     if (doFederatedExport)
-                     {
-                        exportOptions.AddOption("ExportingLinks", selectedConfig.ExportLinkedFiles.ToString());
-                        exportOptions.AddOption("FederatedLinkInfo", linkGuidOptionString);
-                     }
+                     linkExporter = new IFCLinkedDocumentExporter(document, exportOptions);
+                     linkExporter.SetExportOption(selectedConfig.ExportLinkedFiles);
                   }
 
                   bool result = document.Export(path, fileName, exportOptions);
@@ -357,27 +312,17 @@ namespace BIM.IFC.Export.UI
                   // Export links as separate files
                   if (exportSeparateLinks)
                   {
-                     // We can't use the FilterViewId for linked documents, because the
-                     // intermediate code assumes that it is in the exported document.  As such,
-                     // we will pass it in a special place.
-                     ElementId originalFilterViewId = exportOptions.FilterViewId;
-                     if (originalFilterViewId != ElementId.InvalidElementId)
-                     {
-                        exportOptions.AddOption("HostViewId", exportOptions.FilterViewId.ToString());
-                        exportOptions.FilterViewId = ElementId.InvalidElementId;
-                     }
-                     ExportOptionsCache.HostDocument = document;
-                     exportOptions.AddOption("ExportingLinks", LinkedFileExportAs.ExportAsSeparate.ToString());
-                     ExportLinkedDocuments(document, fullName, linkGUIDsCache, linkInstanceTranforms,
-                        exportOptions, originalFilterViewId);
+                     linkExporter.ExportSeparateDocuments(fullName);
                   }
+
+                  string errorMessage = linkExporter?.GetErrors();
 
                   // Show user errors, if any.
                   if (!string.IsNullOrEmpty(errorMessage))
                   {
                      using (Autodesk.Revit.UI.TaskDialog taskDialog = new Autodesk.Revit.UI.TaskDialog(Properties.Resources.IFCExport))
                      {
-                        taskDialog.MainInstruction = string.Format(Properties.Resources.LinkInstanceExportErrorMain, numBadInstances);
+                        taskDialog.MainInstruction = string.Format(Properties.Resources.LinkInstanceExportErrorMain, linkExporter.GetNumberOfBadInstances());
                         taskDialog.MainIcon = Autodesk.Revit.UI.TaskDialogIcon.TaskDialogIconWarning;
                         taskDialog.TitleAutoPrefix = false;
 
@@ -418,349 +363,6 @@ namespace BIM.IFC.Export.UI
                TaskDialogResult result = taskDialog.Show();
             }
          }
-      }
-
-      private string FileNameListToString(IList<string> fileNames)
-      {
-         string fileNameListString = "";
-         foreach (string fileName in fileNames)
-         {
-            if (fileNameListString != "")
-               fileNameListString += "; ";
-            fileNameListString += fileNames;
-         }
-         return fileNameListString;
-      }
-
-      private string ElementIdListToString(IList<ElementId> elementIds)
-      {
-         string elementString = "";
-         foreach (ElementId elementId in elementIds)
-         {
-            if (elementString != "")
-               elementString += ", ";
-            elementString += elementId.ToString();
-         }
-         return elementString;
-      }
-
-      // This modifies an existing string to display an expanded error message to the user.
-      private void AddExpandedStringContent(ref string messageString, string formatString, IList<string> items)
-      {
-         if (messageString != "")
-            messageString += "\n";
-
-         if (items.Count > 0)
-            messageString += string.Format(formatString, FileNameListToString(items));
-      }
-
-      // This modifies an existing string to display an expanded error message to the user.
-      private void AddExpandedElementIdContent(ref string messageString, string formatString, IList<ElementId> items)
-      {
-         if (messageString != "")
-            messageString += "\n";
-
-         if (items.Count > 0)
-            messageString += string.Format(formatString, ElementIdListToString(items));
-      }
-
-      private string GetLinkFileName(Document linkDocument, string linkPathName)
-      {
-         int index = linkPathName.LastIndexOf("\\");
-         if (index <= 0)
-            return linkDocument.Title;
-
-         string linkFileName = linkPathName.Substring(index + 1);
-         // remove the extension
-         index = linkFileName.LastIndexOf('.');
-         if (index > 0)
-            linkFileName = linkFileName.Substring(0, index);
-         return linkFileName;
-      }
-
-      private (IDictionary<RevitLinkInstance, Transform>, string, int) GetLinkedInstanceInfo(
-         IList<RevitLinkInstance> linkInstances)
-      {
-         IDictionary<RevitLinkInstance, Transform> linkedInstanceTransforms =
-            new SortedDictionary<RevitLinkInstance, Transform>(new ElementComparer());
-
-         // We will keep track of the instances we can't export.
-         // Reasons we can't export:
-         // 1. Couldn't create a temporary document for exporting the linked instance.
-         // 2. The document for the linked instance can't be found.
-         // 3. The linked instance is mirrored, non-conformal, or scaled.
-         IList<string> noTempDoc = new List<string>();
-
-         IList<ElementId> nonConformalInst = new List<ElementId>();
-         IList<ElementId> scaledInst = new List<ElementId>();
-         IList<ElementId> instHasReflection = new List<ElementId>();
-
-         int numBadInstances = 0;
-
-         foreach (RevitLinkInstance currRvtLinkInstance in linkInstances)
-         {
-            // Nothing to report if the element itself is null.
-            if (currRvtLinkInstance == null)
-               continue;
-
-            // get the link document and the unit scale
-            Document linkDocument = currRvtLinkInstance.GetLinkDocument();
-            if (linkDocument == null)
-            {
-               // We can't distinguish between unloaded and an error condition, so we
-               // won't get a likely extraneous error.
-               continue;
-            }
-
-            double lengthScaleFactorLink = UnitUtils.ConvertFromInternalUnits(
-               1.0,
-               linkDocument.GetUnits().GetFormatOptions(SpecTypeId.Length).GetUnitTypeId());
-
-            // get the link transform
-            Transform tr = currRvtLinkInstance.GetTransform();
-
-            // We can't handle non-conformal, scaled, or mirrored transforms.
-            ElementId instanceId = currRvtLinkInstance.Id;
-            if (!tr.IsConformal)
-            {
-               nonConformalInst.Add(instanceId);
-               numBadInstances++;
-               continue;
-            }
-
-            if (tr.HasReflection)
-            {
-               instHasReflection.Add(instanceId);
-               numBadInstances++;
-               continue;
-            }
-
-            if (!MathUtil.IsAlmostEqual(tr.Determinant, 1.0))
-            {
-               scaledInst.Add(instanceId);
-               numBadInstances++;
-               continue;
-            }
-
-            // scale the transform origin
-            tr.Origin *= lengthScaleFactorLink;
-
-            linkedInstanceTransforms[currRvtLinkInstance] = tr;
-         }
-
-         // Show user errors, if any.
-         string expandedContent = string.Empty;
-         AddExpandedStringContent(ref expandedContent, Properties.Resources.LinkInstanceExportCantCreateDoc, noTempDoc);
-         AddExpandedElementIdContent(ref expandedContent, Properties.Resources.LinkInstanceExportNonConformal, nonConformalInst);
-         AddExpandedElementIdContent(ref expandedContent, Properties.Resources.LinkInstanceExportScaled, scaledInst);
-         AddExpandedElementIdContent(ref expandedContent, Properties.Resources.LinkInstanceExportHasReflection, instHasReflection);
-
-         return (linkedInstanceTransforms, expandedContent, numBadInstances);
-      }
-
-      /// <summary>
-      /// Checks if element is visible for certain view.
-      /// </summary>
-      /// <param name="element">The element.</param>
-      /// <returns>True if the element is visible, false otherwise.</returns>
-      public bool IsLinkVisible(Element element, View filterView)
-      {
-         if (filterView == null)
-            return true;
-
-         if (element.IsHidden(filterView))
-            return false;
-
-         if (!(element.Category?.get_Visible(filterView) ?? false))
-            return false;
-
-         return filterView.IsElementVisibleInTemporaryViewMode(TemporaryViewMode.TemporaryHideIsolate, element.Id);
-      }
-
-      public void ExportLinkedDocuments(Document document, string fileName,
-         IDictionary<ElementId, string> linkGUIDsCache,
-         IDictionary<RevitLinkInstance, Transform> idToTransform,
-         IFCExportOptions exportOptions, ElementId originalFilterViewId)
-      {
-         // get the extension
-         int index = fileName.LastIndexOf('.');
-         if (index <= 0)
-            return;
-         string sExtension = fileName.Substring(index);
-         fileName = fileName.Substring(0, index);
-
-         // get all the revit link instances
-         IDictionary<string, int> rvtLinkNamesDict = new Dictionary<string, int>();
-         IDictionary<string, List<RevitLinkInstance>> rvtLinkNamesToInstancesDict =
-            new Dictionary<string, List<RevitLinkInstance>>();
-
-         try
-         {
-            View filterView = document.GetElement(originalFilterViewId) as View;
-            foreach (RevitLinkInstance rvtLinkInstance in idToTransform.Keys)
-            {
-               if (!IsLinkVisible(rvtLinkInstance, filterView))
-                  continue;
-
-               // get the link document
-               Document linkDocument = rvtLinkInstance.GetLinkDocument();
-               if (linkDocument == null)
-                  continue;
-
-               // get the link file path and name
-               String linkPathName = "";
-               Parameter originalFileNameParam = linkDocument.ProjectInformation.LookupParameter("Original IFC File Name");
-               if (originalFileNameParam != null && originalFileNameParam.StorageType == StorageType.String)
-                  linkPathName = originalFileNameParam.AsString();
-               else
-                  linkPathName = linkDocument.PathName;
-
-               // get the link file name
-               string linkFileName = GetLinkFileName(linkDocument, linkPathName);
-
-               // add to names count dictionary
-               if (!rvtLinkNamesDict.Keys.Contains(linkFileName))
-                  rvtLinkNamesDict.Add(linkFileName, 0);
-               rvtLinkNamesDict[linkFileName]++;
-
-               // add to names instances dictionary
-               if (!rvtLinkNamesToInstancesDict.Keys.Contains(linkPathName))
-                  rvtLinkNamesToInstancesDict.Add(linkPathName, new List<RevitLinkInstance>());
-               rvtLinkNamesToInstancesDict[linkPathName].Add(rvtLinkInstance);
-            }
-         }
-         catch
-         {
-         }
-
-         foreach (KeyValuePair<string, List<RevitLinkInstance>> linkPathNames in rvtLinkNamesToInstancesDict)
-         {
-            string linkPathName = linkPathNames.Key;
-
-            // get the link instances
-            List<RevitLinkInstance> currRvtLinkInstances = rvtLinkNamesToInstancesDict[linkPathName];
-            IList<string> linkFileNames = new List<string>();
-            IList<Tuple<ElementId, string>> serTransforms = new List<Tuple<ElementId, string>>();
-
-            Document linkDocument = null;
-
-            foreach (RevitLinkInstance currRvtLinkInstance in currRvtLinkInstances)
-            {
-               // Nothing to report if the element itself is null.
-               if (currRvtLinkInstance == null)
-                  continue;
-
-               ElementId instanceId = currRvtLinkInstance.Id;
-
-               // get the link document and the unit scale
-               linkDocument = linkDocument ?? currRvtLinkInstance.GetLinkDocument();
-
-               // get the link file path and name
-               string linkFileName = GetLinkFileName(linkDocument, linkPathName);
-
-               //if link was an IFC file then make a different formating to the file name
-               if ((linkPathName.Length >= 4 && linkPathName.Substring(linkPathName.Length - 4).ToLower() == ".ifc") ||
-                   (linkPathName.Length >= 7 && linkPathName.Substring(linkPathName.Length - 7).ToLower() == ".ifcxml") ||
-                   (linkPathName.Length >= 7 && linkPathName.Substring(linkPathName.Length - 7).ToLower() == ".ifczip"))
-               {
-                  string fName = fileName;
-
-                  //get output path and add to the new file name 
-                  index = fName.LastIndexOf("\\");
-                  if (index > 0)
-                     fName = fName.Substring(0, index + 1);
-                  else
-                     fName = "";
-
-                  //construct IFC file name
-                  linkFileName = fName + linkFileName + "-";
-
-                  //add guid
-                  linkFileName += linkGUIDsCache[instanceId];
-               }
-               else
-               {
-                  // check if there are multiple instances with the same name
-                  bool bMultiple = (rvtLinkNamesDict[linkFileName] > 1);
-
-                  // add the path
-                  linkFileName = fileName + "-" + linkFileName;
-
-                  // add the guid
-                  if (bMultiple)
-                  {
-                     linkFileName += "-";
-                     linkFileName += linkGUIDsCache[instanceId];
-                  }
-               }
-
-               // add the extension
-               linkFileName += sExtension;
-
-               linkFileNames.Add(linkFileName);
-
-               // serialize transform
-               serTransforms.Add(Tuple.Create(instanceId, SerializeTransform(idToTransform[currRvtLinkInstance])));
-            }
-
-            if (linkDocument != null)
-            {
-               // export
-               try
-               {
-                  int numLinkInstancesToExport = linkFileNames.Count;
-                  exportOptions.AddOption("NumberOfExportedLinkInstances", numLinkInstancesToExport.ToString());
-
-                  for (int ind = 0; ind < numLinkInstancesToExport; ind++)
-                  {
-                     string optionName = (ind == 0) ? "ExportLinkId" : "ExportLinkId" + (ind + 1).ToString();
-                     exportOptions.AddOption(optionName, serTransforms[ind].Item1.ToString());
-
-                     optionName = (ind == 0) ? "ExportLinkInstanceTransform" : "ExportLinkInstanceTransform" + (ind + 1).ToString();
-                     exportOptions.AddOption(optionName, serTransforms[ind].Item2);
-
-                     // Don't pass in file name for the first link instance.
-                     if (ind == 0)
-                        continue;
-
-                     optionName = "ExportLinkInstanceFileName" + (ind + 1).ToString();
-                     exportOptions.AddOption(optionName, linkFileNames[ind]);
-                  }
-
-                  // Pass in the first value; the rest will  be in the options.
-                  string path_ = Path.GetDirectoryName(linkFileNames[0]);
-                  string fileName_ = Path.GetFileName(linkFileNames[0]);
-
-                  // Normally, IFC export would need a transaction, even if no permanent
-                  // changes are made.  For linked documents, though, that's handled by the
-                  // export itself.
-                  using (IFCLinkDocumentExportScope scope = new IFCLinkDocumentExportScope(linkDocument))
-                  {
-                     linkDocument.Export(path_, fileName_, exportOptions);
-                  }
-               }
-               catch
-               {
-               }
-            }
-         }
-      }
-
-      public static string SerializeXYZ(XYZ value)
-      {
-         //transform to string
-         return value.ToString();
-      }
-
-      public static string SerializeTransform(Transform tr)
-      {
-         string retVal = string.Empty;
-         //serialize the transform values
-         retVal += SerializeXYZ(tr.Origin) + ";";
-         retVal += SerializeXYZ(tr.BasisX) + ";";
-         retVal += SerializeXYZ(tr.BasisY) + ";";
-         retVal += SerializeXYZ(tr.BasisZ) + ";";
-         return retVal;
       }
    }
 }
